Fix moved all providers inside providers.js
