+++ app/components/forms/EditUser.jsx
@@ -8,8 +8,6 @@ import {
   TextField,
   Typography,
   InputAdornment,
-  ThemeProvider,
-  createTheme,
   FormControl,
   InputLabel,
   Select,
@@ -22,18 +20,6 @@ import { FileUpload } from "@mui/icons-material";
 import InfoIcon from "@mui/icons-material/Info";
 import LoaderStyle from "../helper/LoaderStyle";
 import { useAuth } from "@/context/AuthContext";
-
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: "#00c1a2",
-    },
-    action: {
-      hover: "#00c1a2",
-      selected: "#00c1a2",
-    },
-  },
-});
 
 const EditUser = ({ openEditUserForm, setOpenEditUserForm }) => {
   const handleClose = () => {
@@ -121,43 +107,42 @@ const EditUser = ({ openEditUserForm, setOpenEditUserForm }) => {
   }
 
   return (
-    <ThemeProvider theme={theme}>
-      <Dialog open={openEditUserForm} size={"lg"} className="">
-        <div className="p-6 py-10 ">
-          <div className="flex items-center justify-start gap-2">
-            <IconButton onClick={handleClose} aria-label="back">
-              <ArrowBackIosNewIcon />
-            </IconButton>
-            <h3 className="header-3">Personal Detail:</h3>
+    <Dialog open={openEditUserForm} size={"lg"} className="">
+      <div className="p-6 py-10 ">
+        <div className="flex items-center justify-start gap-2">
+          <IconButton onClick={handleClose} aria-label="back">
+            <ArrowBackIosNewIcon />
+          </IconButton>
+          <h3 className="header-3">Personal Detail:</h3>
+        </div>
+
+        {currentUser ? (
+          currentUser.country === "" || currentUser.bio === "" ? (
+            <p className="text-center text-red-900 py-2">
+              Please complete your personal information
+            </p>
+          ) : null
+        ) : null}
+
+        <form className="flex flex-col gap-5" onSubmit={onSubmit}>
+          <div className="mt-6">
+            <TextField
+              label="Full Name"
+              fullWidth
+              variant="standard"
+              type="text"
+              name="name"
+              value={userData.name}
+              onChange={handleInputChange}
+            />
+            <Typography
+              variant="small"
+              className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
+            >
+              {/* Error messages */}
+            </Typography>
           </div>
-
-          {currentUser ? (
-            currentUser.country === "" || currentUser.bio === "" ? (
-              <p className="text-center text-red-900 py-2">
-                Please complete your personal information
-              </p>
-            ) : null
-          ) : null}
-
-          <form className="flex flex-col gap-5" onSubmit={onSubmit}>
-            <div className="mt-6">
-              <TextField
-                label="Full Name"
-                fullWidth
-                variant="standard"
-                type="text"
-                name="name"
-                value={userData.name}
-                onChange={handleInputChange}
-              />
-              <Typography
-                variant="small"
-                className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
-              >
-                {/* Error messages */}
-              </Typography>
-            </div>
-            {/* <div>
+          {/* <div>
               <TextField
                 label="Email"
                 fullWidth
@@ -175,103 +160,102 @@ const EditUser = ({ openEditUserForm, setOpenEditUserForm }) => {
               </Typography>
             </div> */}
 
-            <div>
-              <TextField
-                label="Bio"
-                fullWidth
-                variant="standard"
-                type="text"
-                name="bio"
-                value={userData.bio}
-                onChange={handleInputChange}
-              />
-              <Typography
-                variant="small"
-                className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
-              >
-                {/* Error messages */}
-              </Typography>
-            </div>
-            <div>
-              <FormControl variant="standard" fullWidth>
-                <InputLabel id="demo-simple-select-label">Country</InputLabel>
-                <Select
-                  labelId="demo-simple-select-label"
-                  id="demo-simple-select"
-                  label="country"
-                  name="country"
-                  value={userData.country}
-                  onChange={handleInputChange}
-                >
-                  {countries.map((country) => {
-                    return (
-                      <MenuItem key={country.label} value={country.label}>
-                        <div className="flex gap-2 items-center">
-                          <span>
-                            <img
-                              className="rounded-none "
-                              loading="lazy"
-                              width="20"
-                              height="10"
-                              src={`https://flagcdn.com/w20/${country.code.toLowerCase()}.png`}
-                              srcSet={`https://flagcdn.com/w40/${country.code.toLowerCase()}.png 2x`}
-                              alt={country.label}
-                            />
-                          </span>
-                          <span>{country.label}</span>
-                        </div>
-                      </MenuItem>
-                    );
-                  })}
-                </Select>
-              </FormControl>
-              <Typography
-                variant="small"
-                className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
-              >
-                {/* Error messages */}
-              </Typography>
-            </div>
-            <div>
-              <TextField
-                fullWidth
-                variant="standard"
-                InputProps={{
-                  startAdornment: (
-                    <InputAdornment position="end">
-                      <FileUpload />
-                    </InputAdornment>
-                  ),
-                }}
-                accept="image/*"
-                id="userImg"
-                name="userImg"
-                type="file"
-                label="User Picture"
-                onChange={handleImageChange}
-                sx={{ input: { cursor: "pointer" } }}
-              />
-              <Typography
-                variant="small"
-                className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
-              >
-                {/* {err} */}
-              </Typography>
-            </div>
-
-            <button
-              className="mt-8 btn-primary w-full "
-              type="submit"
-              variant="filled"
+          <div>
+            <TextField
+              label="Bio"
               fullWidth
+              variant="standard"
+              type="text"
+              name="bio"
+              value={userData.bio}
+              onChange={handleInputChange}
+            />
+            <Typography
+              variant="small"
+              className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
             >
-              Update
-            </button>
-          </form>
-          <p>{err}</p>
-        </div>
-      </Dialog>
-    </ThemeProvider>
+              {/* Error messages */}
+            </Typography>
+          </div>
+          <div>
+            <FormControl variant="standard" fullWidth>
+              <InputLabel id="demo-simple-select-label">Country</InputLabel>
+              <Select
+                labelId="demo-simple-select-label"
+                id="demo-simple-select"
+                label="country"
+                name="country"
+                value={userData.country}
+                onChange={handleInputChange}
+              >
+                {countries.map((country) => {
+                  return (
+                    <MenuItem key={country.label} value={country.label}>
+                      <div className="flex gap-2 items-center">
+                        <span>
+                          <img
+                            className="rounded-none "
+                            loading="lazy"
+                            width="20"
+                            height="10"
+                            src={`https://flagcdn.com/w20/${country.code.toLowerCase()}.png`}
+                            srcSet={`https://flagcdn.com/w40/${country.code.toLowerCase()}.png 2x`}
+                            alt={country.label}
+                          />
+                        </span>
+                        <span>{country.label}</span>
+                      </div>
+                    </MenuItem>
+                  );
+                })}
+              </Select>
+            </FormControl>
+            <Typography
+              variant="small"
+              className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
+            >
+              {/* Error messages */}
+            </Typography>
+          </div>
+          <div>
+            <TextField
+              fullWidth
+              variant="standard"
+              InputProps={{
+                startAdornment: (
+                  <InputAdornment position="end">
+                    <FileUpload />
+                  </InputAdornment>
+                ),
+              }}
+              accept="image/*"
+              id="userImg"
+              name="userImg"
+              type="file"
+              label="User Picture"
+              onChange={handleImageChange}
+              sx={{ input: { cursor: "pointer" } }}
+            />
+            <Typography
+              variant="small"
+              className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
+            >
+              {/* {err} */}
+            </Typography>
+          </div>
+
+          <button
+            className="mt-8 btn-primary w-full "
+            type="submit"
+            variant="filled"
+            fullWidth
+          >
+            Update
+          </button>
+        </form>
+        <p>{err}</p>
+      </div>
+    </Dialog>
   );
 };
