+++ app/components/forms/SignupForm.jsx
@@ -13,8 +13,6 @@ import {
   TextField,
   Typography,
   InputAdornment,
-  ThemeProvider,
-  createTheme,
   FormControl,
   InputLabel,
   Select,
@@ -24,18 +22,6 @@ import {
 import { countries } from "@/data/countries";
 import Link from "next/link";
 import { toast } from "react-toastify";
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: "#00c1a2",
-    },
-    action: {
-      // Customize the autofill background color
-      hover: "#00c1a2", // Replace with your desired color
-      selected: "#00c1a2", // Replace with your desired color
-    },
-  },
-});
 
 const schema = yup
   .object({
@@ -146,173 +132,171 @@ const SignupForm = () => {
         </Link>
       </div>
       <div className="lg:col-span-3 grid gap-2 text-center h-full bg-white lg:rounded-r rounded-t lg:rounded-bl-none order-1 lg:order-2">
-        <ThemeProvider theme={theme}>
-          <div className="grid gap-4 w-full  py-10 px-4">
-            <Container maxWidth="xs">
-              <div className="grid justify-center items-center h-full">
-                <p className="text-center text-gray-900  header-3 px-10 md:px-20">
-                  Create Your Account
-                </p>
-                <form
-                  className="flex flex-col gap-4"
-                  onSubmit={handleSubmit(onSubmit)}
-                >
-                  <div className="mt-6">
-                    <TextField
-                      label="Email"
-                      fullWidth
-                      type="email"
-                      defaultValue="test"
-                      {...register("email")}
-                      variant="standard"
-                    />
-                    <Typography
-                      variant="small"
-                      className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
-                    >
-                      {errors.email && <InfoIcon fontSize="small" />}
-                      {errors.email?.message}
-                    </Typography>
-                  </div>
-                  <div>
-                    <TextField
-                      label="Password"
-                      fullWidth
-                      variant="standard"
-                      type="password"
-                      {...register("password")}
-                    />
-                    <Typography
-                      variant="small"
-                      className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
-                    >
-                      {errors.password && <InfoIcon fontSize="small" />}
-                      {errors.password?.message}
-                    </Typography>
-                  </div>
-                  <div>
-                    <TextField
-                      label="Full Name"
-                      fullWidth
-                      variant="standard"
-                      type="text"
-                      {...register("name")}
-                    />
-                    <Typography
-                      variant="small"
-                      className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
-                    >
-                      {errors.name && <InfoIcon fontSize="small" />}
-                      {errors.name?.message}
-                    </Typography>
-                  </div>
-                  <div>
-                    <TextField
-                      label="Bio"
-                      fullWidth
-                      variant="standard"
-                      type="text"
-                      {...register("bio")}
-                    />
-                    <Typography
-                      variant="small"
-                      className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
-                    >
-                      {errors.bio && <InfoIcon fontSize="small" />}
-                      {errors.bio?.message}
-                    </Typography>
-                  </div>
-                  <div>
-                    <Controller
-                      name="country"
-                      control={control}
-                      render={({ field: { onChange } }) => (
-                        <FormControl variant="standard" fullWidth>
-                          <InputLabel id="demo-simple-select-label">
-                            Country
-                          </InputLabel>
-                          <Select
-                            labelId="demo-simple-select-label"
-                            id="demo-simple-select"
-                            label="country"
-                            onChange={onChange}
-                            defaultValue=""
-                          >
-                            {countries.map((country) => {
-                              return (
-                                <MenuItem
-                                  key={country.label}
-                                  value={country.label}
-                                >
-                                  <div className="flex gap-2 items-center">
-                                    <span>
-                                      <img
-                                        className="rounded-none "
-                                        loading="lazy"
-                                        width="20"
-                                        height="10"
-                                        src={`https://flagcdn.com/w20/${country.code.toLowerCase()}.png`}
-                                      />{" "}
-                                    </span>
-                                    <span>{country.label}</span>
-                                  </div>
-                                </MenuItem>
-                              );
-                            })}
-                          </Select>
-                        </FormControl>
-                      )}
-                    />
-                    <Typography
-                      variant="small"
-                      className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
-                    >
-                      {errors.country && <InfoIcon fontSize="small" />}
-                      {errors.country?.message}
-                    </Typography>
-                  </div>
-                  <div>
-                    <TextField
-                      fullWidth
-                      variant="standard"
-                      InputProps={{
-                        startAdornment: (
-                          <InputAdornment position="end">
-                            <FileUpload />
-                          </InputAdornment>
-                        ),
-                      }}
-                      icon={<FileUpload />}
-                      accept="image/*"
-                      id="userImg"
-                      name="userImg"
-                      type="file"
-                      label="User Picture"
-                      {...register("userImg")}
-                      sx={{ input: { cursor: "pointer" } }}
-                    />
-                    <Typography
-                      variant="small"
-                      className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
-                    >
-                      {errors.userImg && <InfoIcon fontSize="small" />}
-                      {errors.userImg?.message}
-                    </Typography>
-                  </div>
-
-                  <button
-                    className="mt-8 btn-primary w-4/6 self-center "
-                    type="submit"
-                    variant="filled"
+        <div className="grid gap-4 w-full  py-10 px-4">
+          <Container maxWidth="xs">
+            <div className="grid justify-center items-center h-full">
+              <p className="text-center text-gray-900  header-3 px-10 md:px-20">
+                Create Your Account
+              </p>
+              <form
+                className="flex flex-col gap-4"
+                onSubmit={handleSubmit(onSubmit)}
+              >
+                <div className="mt-6">
+                  <TextField
+                    label="Email"
                     fullWidth
+                    type="email"
+                    defaultValue="test"
+                    {...register("email")}
+                    variant="standard"
+                  />
+                  <Typography
+                    variant="small"
+                    className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
                   >
-                    Sign Up
-                  </button>
-                </form>
-              </div>
-            </Container>
-          </div>
-        </ThemeProvider>
+                    {errors.email && <InfoIcon fontSize="small" />}
+                    {errors.email?.message}
+                  </Typography>
+                </div>
+                <div>
+                  <TextField
+                    label="Password"
+                    fullWidth
+                    variant="standard"
+                    type="password"
+                    {...register("password")}
+                  />
+                  <Typography
+                    variant="small"
+                    className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
+                  >
+                    {errors.password && <InfoIcon fontSize="small" />}
+                    {errors.password?.message}
+                  </Typography>
+                </div>
+                <div>
+                  <TextField
+                    label="Full Name"
+                    fullWidth
+                    variant="standard"
+                    type="text"
+                    {...register("name")}
+                  />
+                  <Typography
+                    variant="small"
+                    className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
+                  >
+                    {errors.name && <InfoIcon fontSize="small" />}
+                    {errors.name?.message}
+                  </Typography>
+                </div>
+                <div>
+                  <TextField
+                    label="Bio"
+                    fullWidth
+                    variant="standard"
+                    type="text"
+                    {...register("bio")}
+                  />
+                  <Typography
+                    variant="small"
+                    className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
+                  >
+                    {errors.bio && <InfoIcon fontSize="small" />}
+                    {errors.bio?.message}
+                  </Typography>
+                </div>
+                <div>
+                  <Controller
+                    name="country"
+                    control={control}
+                    render={({ field: { onChange } }) => (
+                      <FormControl variant="standard" fullWidth>
+                        <InputLabel id="demo-simple-select-label">
+                          Country
+                        </InputLabel>
+                        <Select
+                          labelId="demo-simple-select-label"
+                          id="demo-simple-select"
+                          label="country"
+                          onChange={onChange}
+                          defaultValue=""
+                        >
+                          {countries.map((country) => {
+                            return (
+                              <MenuItem
+                                key={country.label}
+                                value={country.label}
+                              >
+                                <div className="flex gap-2 items-center">
+                                  <span>
+                                    <img
+                                      className="rounded-none "
+                                      loading="lazy"
+                                      width="20"
+                                      height="10"
+                                      src={`https://flagcdn.com/w20/${country.code.toLowerCase()}.png`}
+                                    />{" "}
+                                  </span>
+                                  <span>{country.label}</span>
+                                </div>
+                              </MenuItem>
+                            );
+                          })}
+                        </Select>
+                      </FormControl>
+                    )}
+                  />
+                  <Typography
+                    variant="small"
+                    className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
+                  >
+                    {errors.country && <InfoIcon fontSize="small" />}
+                    {errors.country?.message}
+                  </Typography>
+                </div>
+                <div>
+                  <TextField
+                    fullWidth
+                    variant="standard"
+                    InputProps={{
+                      startAdornment: (
+                        <InputAdornment position="end">
+                          <FileUpload />
+                        </InputAdornment>
+                      ),
+                    }}
+                    icon={<FileUpload />}
+                    accept="image/*"
+                    id="userImg"
+                    name="userImg"
+                    type="file"
+                    label="User Picture"
+                    {...register("userImg")}
+                    sx={{ input: { cursor: "pointer" } }}
+                  />
+                  <Typography
+                    variant="small"
+                    className="flex items-center gap-1 font-normal mt-2 text-red-800 mb-4"
+                  >
+                    {errors.userImg && <InfoIcon fontSize="small" />}
+                    {errors.userImg?.message}
+                  </Typography>
+                </div>
+
+                <button
+                  className="mt-8 btn-primary w-4/6 self-center "
+                  type="submit"
+                  variant="filled"
+                  fullWidth
+                >
+                  Sign Up
+                </button>
+              </form>
+            </div>
+          </Container>
+        </div>
       </div>
     </div>
   );
